requirements.txt
Django==1.4.13
MySQL-python==1.2.5
django-admin-bootstrapped==1.6.5
django-extensions==1.2.5
django-filter==0.7
django-oauth2-provider==0.2.6.1
djangorestframework==2.3.5
djangorestframework-csv==1.3.3
geoip2==0.5.0
ipaddr==2.1.11
mongoengine==0.8.7
pep8==1.4.6
pycountry==1.6
pymongo==2.6.3
python-dateutil==2.1
python-memcached==1.53
python-termstyle==0.1.10
simplejson==3.3.3
urllib3==1.8.2
virtualenv==1.11.4
wsgiref==0.1.2
django-cors-headers
djangorestframework-jwt
google-api-python-client
httplib2
python-gflags